pyproject.toml
[project]
name = "jobs-china"
version = "0.1.0"
description = "China Job Market Visualizer - AI exposure scoring for Chinese occupations"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true